refactor: moved login from Individual components to Context.
