+++ src/utils/index.js
@@ -4,6 +4,7 @@ export { priceFilter } from "./priceFilter";
 export { ratingFilter } from "./ratingFilter";
 export { categoryFilter } from "./categoryFilter";
 export { productSearchFilter } from "./productSearchFilter";
-export { getCartSummary } from "../utils/getCartSummary";
-export { getDiscountPercentage } from "../utils/getDiscountPercentage";
-export { getDate } from "../utils/getDate";
+export { getCartSummary } from "./getCartSummary";
+export { getDiscountPercentage } from "./getDiscountPercentage";
+export { getDate } from "./getDate";
+export { productImagesSlider, similarProductSlider } from "./sliderSettings";
